# docs/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SecureVOTE API",
    "description": "Educational/research prototype of an embedded electronic-voting and election-integrity pipeline. This is NOT production election infrastructure. See /docs for the interactive API explorer.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/elections": {
      "get": {
        "tags": [
          "elections"
        ],
        "summary": "List Elections",
        "description": "List all elections.",
        "operationId": "list_elections_api_elections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectionListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "elections"
        ],
        "summary": "Create Election",
        "description": "Create a new election. Requires ADMIN role.",
        "operationId": "create_election_api_elections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/elections/{election_id}": {
      "get": {
        "tags": [
          "elections"
        ],
        "summary": "Get Election",
        "description": "Get election details with candidates and device count.",
        "operationId": "get_election_api_elections__election_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/elections/{election_id}/candidates": {
      "post": {
        "tags": [
          "elections"
        ],
        "summary": "Add Candidates",
        "description": "Add candidates to an election. Requires ADMIN role.",
        "operationId": "add_candidates_api_elections__election_id__candidates_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateBatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateResponse"
                  },
                  "title": "Response Add Candidates Api Elections  Election Id  Candidates Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "elections"
        ],
        "summary": "Get Candidates",
        "description": "Get candidates for an election.",
        "operationId": "get_candidates_api_elections__election_id__candidates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateResponse"
                  },
                  "title": "Response Get Candidates Api Elections  Election Id  Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/elections/{election_id}/results": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Results",
        "description": "Get election results with full independent verification.\n\nThis endpoint does NOT return cached results \u00e2\u20ac\u201d it independently\nrecomputes tallies, verifies the audit chain, and runs reconciliation\nevery time it is called.",
        "operationId": "get_results_api_elections__election_id__results_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/elections/{election_id}/verify": {
      "post": {
        "tags": [
          "results"
        ],
        "summary": "Run Verification",
        "description": "Run full independent verification and generate/update result manifest.\nRequires ADMIN or AUDITOR role.",
        "operationId": "run_verification_api_elections__election_id__verify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/transparency/elections/{election_id}": {
      "get": {
        "tags": [
          "Transparency"
        ],
        "summary": "Get Transparency Overview",
        "operationId": "get_transparency_overview_api_transparency_elections__election_id__get",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransparencyOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transparency/elections/{election_id}/candidates": {
      "get": {
        "tags": [
          "Transparency"
        ],
        "summary": "Get Transparency Candidates",
        "operationId": "get_transparency_candidates_api_transparency_elections__election_id__candidates_get",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateResponse"
                  },
                  "title": "Response Get Transparency Candidates Api Transparency Elections  Election Id  Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transparency/elections/{election_id}/results": {
      "get": {
        "tags": [
          "Transparency"
        ],
        "summary": "Get Transparency Results",
        "operationId": "get_transparency_results_api_transparency_elections__election_id__results_get",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateBatchCreate": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Candidates"
          }
        },
        "type": "object",
        "required": [
          "candidates"
        ],
        "title": "CandidateBatchCreate"
      },
      "CandidateCreate": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^C\\d{3}$",
            "title": "Id",
            "examples": [
              "C001"
            ]
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "party": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Party"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "position": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "position"
        ],
        "title": "CandidateCreate"
      },
...
      "CandidateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "election_id": {
            "type": "string",
            "title": "Election Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "party": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Party"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "election_id",
          "name",
          "party",
          "symbol",
          "position",
          "created_at"
        ],
        "title": "CandidateResponse"
      },
...
      "DeviceResult": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "device_name": {
            "type": "string",
            "title": "Device Name"
          },
          "ballot_count": {
            "type": "integer",
            "title": "Ballot Count"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "device_name",
          "ballot_count"
        ],
        "title": "DeviceResult"
      },
...
      "ElectionCreate": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^EV-\\d{4}-\\d{3}$",
            "title": "Id",
            "examples": [
              "EV-2026-001"
            ]
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ElectionCreate"
      },
...
      "ElectionListResponse": {
        "properties": {
          "elections": {
            "items": {
              "$ref": "#/components/schemas/ElectionResponse"
            },
            "type": "array",
            "title": "Elections"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "elections",
          "total"
        ],
        "title": "ElectionListResponse"
      },
      "ElectionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "configuration_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration Hash"
          },
          "total_ballots": {
            "type": "integer",
            "title": "Total Ballots"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "configured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configured At"
          },
          "locked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked At"
          },
          "opened_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateResponse"
            },
            "type": "array",
            "title": "Candidates",
            "default": []
          },
          "device_count": {
            "type": "integer",
            "title": "Device Count",
            "default": 0
          },
          "ballot_count": {
            "type": "integer",
            "title": "Ballot Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "state",
          "configuration_hash",
          "total_ballots",
          "created_at",
          "configured_at",
          "locked_at",
          "opened_at",
          "closed_at",
          "published_at"
        ],
        "title": "ElectionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResultManifestResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "election_id": {
            "type": "string",
            "title": "Election Id"
          },
          "total_ballots": {
            "type": "integer",
            "title": "Total Ballots"
          },
          "candidate_totals": {
            "type": "string",
            "title": "Candidate Totals"
          },
          "device_totals": {
            "type": "string",
            "title": "Device Totals"
          },
          "reconciliation_status": {
            "type": "string",
            "title": "Reconciliation Status"
          },
          "audit_chain_status": {
            "type": "string",
            "title": "Audit Chain Status"
          },
          "configuration_hash": {
            "type": "string",
            "title": "Configuration Hash"
          },
          "manifest_hash": {
            "type": "string",
            "title": "Manifest Hash"
          },
          "digital_signature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digital Signature"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "verified_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified By"
          },
          "candidate_results": {
            "items": {
              "$ref": "#/components/schemas/CandidateResult"
            },
            "type": "array",
            "title": "Candidate Results",
            "default": []
          },
          "device_results": {
            "items": {
              "$ref": "#/components/schemas/DeviceResult"
            },
            "type": "array",
            "title": "Device Results",
            "default": []
          },
          "reconciliation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReconciliationResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "election_id",
          "total_ballots",
          "candidate_totals",
          "device_totals",
          "reconciliation_status",
          "audit_chain_status",
          "configuration_hash",
          "manifest_hash",
          "digital_signature",
          "generated_at",
          "verified_at",
          "verified_by"
        ],
        "title": "ResultManifestResponse"
      },
...
      "TransparencyOverview": {
        "properties": {
          "election_id": {
            "type": "string",
            "title": "Election Id"
          },
          "election_name": {
            "type": "string",
            "title": "Election Name"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "total_ballots": {
            "type": "integer",
            "title": "Total Ballots"
          },
          "device_count": {
            "type": "integer",
            "title": "Device Count"
          },
          "polling_station_count": {
            "type": "integer",
            "title": "Polling Station Count"
          },
          "candidate_count": {
            "type": "integer",
            "title": "Candidate Count"
          },
          "registered_voters": {
            "type": "integer",
            "title": "Registered Voters"
          },
          "turnout_percentage": {
            "type": "number",
            "title": "Turnout Percentage"
          },
          "audit_chain_status": {
            "type": "string",
            "title": "Audit Chain Status"
          },
          "reconciliation_status": {
            "type": "string",
            "title": "Reconciliation Status"
          },
          "manifest_status": {
            "type": "string",
            "title": "Manifest Status"
          },
          "notice": {
            "type": "string",
            "title": "Notice",
            "default": "Aggregate data only - no individual voter information disclosed"
          }
        },
        "type": "object",
        "required": [
          "election_id",
          "election_name",
          "state",
          "total_ballots",
          "device_count",
          "polling_station_count",
          "candidate_count",
          "registered_voters",
          "turnout_percentage",
          "audit_chain_status",
          "reconciliation_status",
          "manifest_status"
        ],
        "title": "TransparencyOverview"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationResponse": {
        "properties": {
          "election_id": {
            "type": "string",
            "title": "Election Id"
          },
          "config_hash_valid": {
            "type": "boolean",
            "title": "Config Hash Valid"
          },
          "audit_chain_intact": {
            "type": "boolean",
            "title": "Audit Chain Intact"
          },
          "reconciliation_passed": {
            "type": "boolean",
            "title": "Reconciliation Passed"
          },
          "tally_independently_verified": {
            "type": "boolean",
            "title": "Tally Independently Verified"
          },
          "overall_status": {
            "type": "string",
            "title": "Overall Status"
          },
          "manifest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultManifestResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "details": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Details",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "election_id",
          "config_hash_valid",
          "audit_chain_intact",
          "reconciliation_passed",
          "tally_independently_verified",
          "overall_status"
        ],
        "title": "VerificationResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}